+++ src/models/types/users.types.ts
@@ -1,0 +1,6 @@
+import { Document } from "mongoose";
+
+export interface IUser extends Document {
+  email: string;
+  password: string;
+}
